Upgrading webpack dev env to latest version
