+++ dev/webpack/config.js
@@ -1,29 +1,74 @@
 const path = require("path")
-const config = require("../../webpack.config")
-const webpack = require("webpack")
+const ForkTsCheckerWebpackPlugin = require("fork-ts-checker-webpack-plugin")
+const convertPathsToAliases = require("convert-tsconfig-paths-to-webpack-aliases")
+    .default
 const chalk = require("chalk")
+const tsconfig = require("../../tsconfig.json")
+
+const tsLoader = {
+    loader: "ts-loader",
+    options: { transpileOnly: true },
+}
 
 const DEV_SERVER_PORT = 9990
 
-console.log(chalk.bold.green(`\nRunning at: http://0.0.0.0:${DEV_SERVER_PORT}/\n`))
+console.log(
+    chalk.bold.green(`\nRunning at: http://0.0.0.0:${DEV_SERVER_PORT}/\n`)
+)
 
-config.entry = {
-    framer: [
-        "react-dev-utils/webpackHotDevClient",
-        path.join(__dirname, "..", "..", "src", "index"),
-        path.join(__dirname, "index"),
-    ],
+module.exports = {
+    mode: "development",
+    target: "web",
+    entry: {
+        framer: [
+            path.join(__dirname, "..", "..", "src", "index"),
+            path.join(__dirname, "index"),
+        ],
+    },
+    output: {
+        path: path.join(__dirname, "build"),
+        filename: "[name].debug.js",
+        library: "Framer",
+        libraryTarget: "umd",
+    },
+    devtool: "eval-cheap-source-map",
+    module: {
+        rules: [
+            {
+                test: /\.ts(x?)$/,
+                exclude: [/__tests__/, /node_modules/],
+                use: ["cache-loader", tsLoader],
+            },
+        ],
+    },
+    resolve: {
+        modules: ["node_modules"],
+        extensions: [".ts", ".tsx", ".js", ".json"],
+        alias: convertPathsToAliases(tsconfig),
+        fallback: {
+            path: require.resolve("path-browserify"),
+        },
+    },
+    externals: {
+        react: {
+            root: "React",
+            commonjs2: "react",
+            commonjs: "react",
+            amd: "react",
+        },
+        "react-dom": {
+            root: "ReactDOM",
+            commonjs2: "ReactDOM",
+            commonjs: "ReactDOM",
+            amd: "ReactDOM",
+        },
+    },
+    devServer: {
+        contentBase: path.join(__dirname),
+        hot: true,
+        inline: false,
+        stats: "errors-only",
+        port: DEV_SERVER_PORT,
+    },
+    plugins: [new ForkTsCheckerWebpackPlugin()],
 }
-
-config.devServer = {
-    contentBase: path.join(__dirname),
-    hot: false,
-    inline: false,
-    // quiet: true,
-    // open: true,
-    // openPage: "/",
-    stats: "errors-only",
-    port: DEV_SERVER_PORT,
-}
-
-Object.assign(exports, config)
